requirements.txt
django==4.2.6
django-admin-tools
django-crispy-forms
django-debug-toolbar
django-extensions
django_cleanup
django_bootstrap_icons
crispy-bootstrap5
mysqlclient
paho-mqtt
sox
Pillow
astropy
humanize
tzlocal
numpy
scipy
matplotlib
tflite-runtime
tuya-iot-py-sdk
